requirements.txt
flask
gunicorn
flask_wtf
pyyaml
loguru
numpy
torch
transformers
boto3